Split the inventory strategy route into section components
Move each page section into its own component under the feature's
components directory, with the shared formatters and scenario selection
helpers in modules the sections import. The horizon curve owns the capital
cycle inputs and the forecast grading owns its horizon toggle, so the route
keeps only the loader, the action, the scenario selections, and the layout.
No rendered output changes.

Closes #1
