+++ app/features/inventory-strategy/components/MetricCard.tsx
@@ -1,0 +1,27 @@
+import { Card, CardContent, Typography } from "@mui/material";
+
+export function MetricCard({
+  label,
+  value,
+  detail,
+}: {
+  label: string;
+  value: string;
+  detail: string;
+}) {
+  return (
+    <Card variant="outlined">
+      <CardContent>
+        <Typography variant="body2" color="text.secondary">
+          {label}
+        </Typography>
+        <Typography variant="h5" sx={{ my: 0.5 }}>
+          {value}
+        </Typography>
+        <Typography variant="caption" color="text.secondary">
+          {detail}
+        </Typography>
+      </CardContent>
+    </Card>
+  );
+}
